+++ main-program/Makefile
@@ -1,7 +1,7 @@
 SHELL = /bin/sh
 
 # Build tools and flags
-OBJS = src/main.cpp src/opengl.cpp src/socketdev.cpp src/glad.cpp
+OBJS = src/main.cpp src/scene.cpp src/socketdev.cpp src/glad.cpp
 CFLAGS = -lglfw3 -lX11 -lGL -lpthread -lXrandr -lXi -ldl
 CFLAGSI = -ljson-c
 CC = g++
