+++ main-program/src/main.cpp
@@ -1,7 +1,10 @@
-#include "opengl.h"
+#include "scene.h"
 #include "socketdev.h"
 
-
+#define ISVALIDSOCKET(s) ((s) >= 0)
+#define CLOSESOCKET(s) close(s)
+#define SOCKET int
+#define GETSOCKETERRNO() (errno)
 
 //#define DEBUG
 
@@ -10,26 +13,26 @@ int main(int argc, char* argv[]){
         char *buffer_pointer;
 
         socket_programming socket(argc, argv);
-        GL gl;
+        Scene scene;
         /* initiate the communication with the simulation */
         socket.init_socket();
         /* initiate the window */
-        gl.init_window();
+        scene.init_glfw();
         /* socket connections */
         socket.print_addr();
         socket.create();
         socket.make_connection();
         /*assign some value to memory and render the first frame */
         buffer_pointer = socket.run();
-        gl.json_parser(buffer_pointer);
-        gl.compile_shader();
-        gl.draw();
+        // gl.json_parser(buffer_pointer);
+        scene.compile_shader();
+        // gl.draw();
         while(1){
             buffer_pointer = socket.run();
-            gl.json_parser(buffer_pointer);
-            gl.draw();
+            // gl.json_parser(buffer_pointer);
+            scene.draw();
         }
-        socket.cleanup();
+        // socket.cleanup();
     }
     catch(...){
         std::cout<<"Error occured..."<<std::endl;
