+++ main-program/src/shader/basic.vert
@@ -1,7 +1,8 @@
 #version 450
+lets make an obvious error!!
 uniform mat4 MVP;
-layout(location=0) in vec2 vPos;
-layout(location=1) in vec3 vCol;
+layout(location=0) in vec3 vCol;
+layout(location=1) in vec2 vPos;
 out vec3 color;
 void main()
 {
